pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "majordomo-concierge-v2"
version = "0.1.0"
description = "Majordomo Concierge V2 - multi-agent personal AI assistant"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.19.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "chromadb>=0.4.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["src"]
